Divide DCAT-US harvesters for CKAN 2.8 or CKAN 2.3
